feat: Add Player Upgrade System
- 플레이어가 Gem을 4개 이상 모으면 Upgrade 창 띄우기
- 플레이어가 선택한 것 Upgrade 시켜주기

+++ Assets/01.Scripts/Player/CannedTuna.cs
@@ -9,7 +9,7 @@ public class CannedTuna : MonoBehaviour
     float timer;
     float lifeTime = 3f;
 
-    int damage;
+    float damage;
 
     void Start()
     {
@@ -36,7 +36,7 @@ public class CannedTuna : MonoBehaviour
         rb.linearVelocity = transform.right * speed;
     }
 
-    public void SetDamage(int dmg)
+    public void SetDamage(float dmg)
     {
         damage = dmg;
     }
@@ -45,7 +45,6 @@ public class CannedTuna : MonoBehaviour
     {
         if (collision.gameObject.layer == LayerMask.NameToLayer("Enemy"))
         {
-            //몬스터 데미지 입히기
             collision.gameObject.GetComponent<EnemyController>().TakeDamage(damage);
             ReturnPool();
         }
